pygstc: Fix socket being wrongly configure to non-blocking
Due to what seems to be a refactoring error, the timeout argument was
being wrongly used. It was documented that 0 mean blocking mode, but
underneath this value was not appropriately used among select and
socket. Pygstc uses a single socket to make requests. Given this,
waiting on select is equivalent to waiting on this socket.

In favor of simplicity we now use the socket module timeout standard.